+++ examples/vocab/pytext_vocab.py
@@ -2,7 +2,8 @@ from collections import OrderedDict
 
 from fairseq.data.dictionary import Dictionary
 import torch
-from torchtext.experimental.vocab import vocab, Vocab
+from torchtext.experimental.vocab import vocab
+from torchtext.vocab import Vocab
 from typing import Dict, List, Optional
