+++ test/test_vocab.py
