+++ torchtext/experimental/vocab.py
@@ -1,8 +1,6 @@
 import logging
-from typing import Dict, List, Optional
 from collections import Counter, OrderedDict
-import torch
-import torch.nn as nn
+from torchtext.vocab import Vocab
 from torchtext._torchtext import (
     Vocab as VocabPybind,
     _load_vocab_from_file,
@@ -13,7 +11,6 @@ __all__ = [
     'build_vocab_from_text_file',
     'load_vocab_from_file',
     'vocab',
-    'Vocab',
 ]
 logger = logging.getLogger(__name__)
 
@@ -138,174 +135,3 @@ def vocab(ordered_dict, min_freq=1, unk_token='<unk>'):
     if unk_token not in tokens:
         tokens.insert(0, unk_token)
     return Vocab(VocabPybind(tokens, None))
-
-
-class Vocab(nn.Module):
-    __jit_unused_properties__ = ["is_jitable"]
-    r"""Creates a vocab object which maps tokens to indices.
-
-    Args:
-        vocab (torch.classes.torchtext.Vocab or torchtext._torchtext.Vocab): a cpp vocab object.
-    """
-
-    def __init__(self, vocab):
-        super(Vocab, self).__init__()
-        self.vocab = vocab
-
-    @property
-    def is_jitable(self):
-        return not isinstance(self.vocab, VocabPybind)
-
-    @torch.jit.export
-    def forward(self, tokens: List[str]) -> List[int]:
-        r"""Calls the `lookup_indices` method
-
-        Args:
-            tokens (List[str]): a list of tokens used to lookup their corresponding `indices`.
-
-        Returns:
-            indices (List[int]): the 'indices` associated with a list of tokens`.
-        """
-        return self.vocab.lookup_indices(tokens)
-
-    @torch.jit.export
-    def __len__(self) -> int:
-        r"""
-        Returns:
-            length (int): the length of the vocab
-        """
-        return len(self.vocab)
-
-    @torch.jit.export
-    def __contains__(self, token: str) -> bool:
-        r"""
-        Args:
-            token (str): the token for which to check the membership
-
-        Returns:
-            membership (bool): whether the token is member of vocab or not
-        """
-        return self.vocab.__contains__(token)
-
-    @torch.jit.export
-    def __getitem__(self, token: str) -> int:
-        r"""
-        Args:
-            token (str): the token used to lookup the corresponding index.
-
-        Returns:
-            index (int): the index corresponding to the associated token.
-        """
-        return self.vocab[token]
-
-    @torch.jit.export
-    def set_default_index(self, index: int) -> None:
-        r"""
-        Args:
-            index: Value of default index. This index will be returned when OOV token is queried
-        """
-        self.vocab.set_default_index(index)
-
-    @torch.jit.export
-    def get_default_index(self) -> Optional[int]:
-        r"""
-        Returns:
-            index (optional[int]): Value of default index if it is set.
-        """
-        return self.vocab.get_default_index()
-
-    @torch.jit.export
-    def reassign_token(self, token: str, index: int) -> None:
-        r"""
-        Args:
-            token (str): the token used to lookup the corresponding index.
-            index (int): the index corresponding to the associated token.
-        Raises:
-            RuntimeError: If `index` is not range [0,Vocab.size()) or if token is not present in Vocab
-        """
-        self.vocab.reassign_token(token, index)
-
-    @torch.jit.export
-    def insert_token(self, token: str, index: int) -> None:
-        r"""
-        Args:
-            token (str): the token used to lookup the corresponding index.
-            index (int): the index corresponding to the associated token.
-        Raises:
-            RuntimeError: if `index` not between [0, Vocab.size()] or if token already exists in the vocab.
-        """
-        self.vocab.insert_token(token, index)
-
-    @torch.jit.export
-    def append_token(self, token: str) -> None:
-        r"""
-        Args:
-            token (str): the token used to lookup the corresponding index.
-
-        Raises:
-            RuntimeError: if token already exists in the vocab
-        """
-        self.vocab.append_token(token)
-
-    @torch.jit.export
-    def lookup_token(self, index: int) -> str:
-        r"""
-        Args:
-            index (int): the index corresponding to the associated token.
-
-        Returns:
-            token (str): the token used to lookup the corresponding index.
-
-        Raises:
-            RuntimeError: if `index` not between [0, itos.size()].
-        """
-        return self.vocab.lookup_token(index)
-
-    @torch.jit.export
-    def lookup_tokens(self, indices: List[int]) -> List[str]:
-        r"""
-        Args:
-            indices (List[int]): the `indices` used to lookup their corresponding`tokens`.
-
-        Returns:
-            tokens (List[str]): the `tokens` associated with `indices`.
-
-        Raises:
-            RuntimeError: if an index within `indices` is not between [0, itos.size()].
-        """
-        return self.vocab.lookup_tokens(indices)
-
-    @torch.jit.export
-    def lookup_indices(self, tokens: List[str]) -> List[int]:
-        r"""
-        Args:
-            tokens (List[str]): the tokens used to lookup their corresponding `indices`.
-
-        Returns:
-            indices (List[int]): the 'indices` associated with `tokens`.
-        """
-        return self.vocab.lookup_indices(tokens)
-
-    @torch.jit.export
-    def get_stoi(self) -> Dict[str, int]:
-        r"""
-        Returns:
-            stoi (dict): dictionary mapping tokens to indices.
-        """
-        return self.vocab.get_stoi()
-
-    @torch.jit.export
-    def get_itos(self) -> List[str]:
-        r"""
-        Returns:
-            itos (dict): dictionary mapping indices to tokens.
-        """
-        return self.vocab.get_itos()
-
-    def __prepare_scriptable__(self):
-        r"""Return a JITable Vocab.
-        """
-        if not self.is_jitable:
-            cpp_vocab = torch.classes.torchtext.Vocab(self.vocab.itos_, self.vocab.default_index_)
-            return Vocab(cpp_vocab)
-        return self
